Apparent fix to the menu problem
The crux seems to be to *not* include PointerMotionHintMask in the
argument to XGrabPointer.  The problem is that if PointerMotionHintMask
is set, then the server is permitted to delay sending pointer motion
events until something other than pointer motion happens.

https://tronche.com/gui/x/xlib/events/keyboard-pointer/keyboard-pointer.html
    If PointerMotionHintMask is selected in combination with one or more of
    the above masks, the X server is free to send only one MotionNotify
    event (with the is_hint member of the XPointerMovedEvent structure set
    to NotifyHint) to the client for the event window, until either the key
    or button state changes, the pointer leaves the event window, or the
    client calls XQueryPointer() or XGetMotionEvents().

Apparently, the X server doesn't actually delay motion events except in
certain circumstances, and the menu problem case is one of those special
cases.