An improvement to GetColor()
There is a function called GetColor().  I don't know why, but its
behavior depends on Scr->FirstTime:  If that flag is true, GetColor does
the lookup it is supposed to do.  If that flag is false, GetColor does
nothing.

The problem is that there are many calls of GetColor where the caller
wants the lookup done even if Scr->FirstTime is false, leading to many
instances of the idiom:

    int save;

    save = Scr->FirstTime;
        Scr->FirstTime = TRUE;

    GetColor(COLOR, &tmp->highlight.fore, fore);
        GetColor(COLOR, &tmp->highlight.back, back);

    Scr->FirstTime = save;

I've made a change that separates the processing part of GetColor into
GetColorAlways, which is not sensitive to Scr->FirstTime.  GetColor
then becomes a wrapper function that tests Scr->FirstTime and then
conditionally calls GetColorAlways.  The above idiom is then replaced by
two calls of GetColorAlways.

The code seems to work for me, but I do not have a good idea what the
code does.  Could somebody check it over?

(Beware that to regenerate gram.c from gram.y, you need to do "rm gram.c
; make".)

Dale